Use the modern tomli/tomllib to parse TOML files
Replace the unmaintained `toml` package with the modern combination
of built-in `tomllib` in Python 3.11+, and its drop-in replacement
`tomli` for older Python versions.  This provides full and correct
support for TOML 1.0 that `toml` does not provide.  Unfortunately, they
do not support overriding dict class but that does not seem to be
a major problem, given that TOML tables are not ordered according
to the specification.